+++ src/Instances.js
@@ -1,6 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import _ from 'lodash';
+import {
+  omit,
+  keyBy,
+  get,
+  template,
+} from 'lodash';
 import {
   injectIntl,
   intlShape,
@@ -62,14 +67,14 @@ const filterConfig = [
 ];
 
 const searchableIndexes = [
-  { label: 'ui-inventory.search.all', value: 'all', queryTemplate: 'title="%{query.query}" or contributors adj "\\"name\\": \\"%{query.query}\\"" or identifiers adj "\\"value\\": \\"%{query.query}\\""' },
+  { label: 'ui-inventory.search.all', value: 'all', queryTemplate: 'title="%{query.query}" or contributors =/@name "%{query.query}" or identifiers =/@value "%{query.query}"' },
   { label: 'ui-inventory.barcode', value: 'item.barcode', queryTemplate: 'item.barcode=="%{query.query}"' },
   { label: 'ui-inventory.instanceId', value: 'id', queryTemplate: 'id="%{query.query}"' },
   { label: 'ui-inventory.title', value: 'title', queryTemplate: 'title="%{query.query}"' },
-  { label: 'ui-inventory.identifier', value: 'identifier', queryTemplate: 'identifiers adj "\\"value\\": \\"%{query.query}\\""' },
-  { label: 'ui-inventory.isbn', prefix: '- ', value: 'isbn', queryTemplate: 'identifiers == "*\\"value\\": \\"%{query.query}\\", \\"identifierTypeId\\": \\"<%= identifierTypeId %>\\""' },
-  { label: 'ui-inventory.issn', prefix: '- ', value: 'issn', queryTemplate: 'identifiers == "*\\"value\\": \\"%{query.query}\\", \\"identifierTypeId\\": \\"<%= identifierTypeId %>\\""' },
-  { label: 'ui-inventory.contributor', value: 'contributor', queryTemplate: 'contributors adj "\\"name\\": \\"%{query.query}\\""' },
+  { label: 'ui-inventory.identifier', value: 'identifier', queryTemplate: 'identifiers =/@value "%{query.query}"' },
+  { label: 'ui-inventory.isbn', prefix: '- ', value: 'isbn', queryTemplate: 'identifiers =/@value/@identifierTypeId="<%= identifierTypeId %>" "%{query.query}"' },
+  { label: 'ui-inventory.issn', prefix: '- ', value: 'issn', queryTemplate: 'identifiers =/@value/@identifierTypeId="<%= identifierTypeId %>" "%{query.query}"' },
+  { label: 'ui-inventory.contributor', value: 'contributor', queryTemplate: 'contributors =/@name "%{query.query}"' },
   { label: 'ui-inventory.subject', value: 'subject', queryTemplate: 'subjects="%{query.query}"' },
 ];
 
@@ -112,7 +117,7 @@ class Instances extends React.Component {
               const identifierType = resourceData.identifier_types.records.find(type => type.name.toLowerCase() === queryIndex);
               const identifierTypeId = identifierType ? identifierType.id : 'identifier-type-not-found';
 
-              queryTemplate = _.template(searchableIndex.queryTemplate)({ identifierTypeId });
+              queryTemplate = template(searchableIndex.queryTemplate)({ identifierTypeId });
             } else {
               queryTemplate = searchableIndex.queryTemplate;
             }
@@ -189,6 +194,15 @@ class Instances extends React.Component {
       type: 'okapi',
       records: 'issuanceModes',
       path: 'modes-of-issuance?limit=1000&query=cql.allRecords=1 sortby name',
+    },
+    instanceNoteTypes: {
+      type: 'okapi',
+      path: 'instance-note-types',
+      params: {
+        query: 'cql.allRecords=1 sortby name',
+        limit: '1000',
+      },
+      records: 'instanceNoteTypes',
     },
     electronicAccessRelationships: {
       type: 'okapi',
@@ -287,7 +301,7 @@ class Instances extends React.Component {
   }
 
   copyInstance(instance) {
-    this.setState({ copiedInstance: _.omit(instance, ['id', 'hrid']) });
+    this.setState({ copiedInstance: omit(instance, ['id', 'hrid']) });
     this.props.updateLocation({ layer: 'create' });
   }
 
@@ -321,6 +335,7 @@ class Instances extends React.Component {
       || !resources.instanceStatuses || !resources.instanceStatuses.hasLoaded
       || !resources.modesOfIssuance || !resources.modesOfIssuance.hasLoaded
       || !resources.electronicAccessRelationships || !resources.electronicAccessRelationships.hasLoaded
+      || !resources.instanceNoteTypes || !resources.instanceNoteTypes.hasLoaded
       || !resources.statisticalCodeTypes || !resources.statisticalCodeTypes.hasLoaded
       || !resources.statisticalCodes || !resources.statisticalCodes.hasLoaded
       || !resources.illPolicies || !resources.illPolicies.hasLoaded
@@ -340,6 +355,7 @@ class Instances extends React.Component {
     const instanceStatuses = (resources.instanceStatuses || emptyObj).records || emptyArr;
     const modesOfIssuance = (resources.modesOfIssuance || emptyObj).records || emptyArr;
     const electronicAccessRelationships = (resources.electronicAccessRelationships || emptyObj).records || emptyArr;
+    const instanceNoteTypes = resources.instanceNoteTypes.records;
     const statisticalCodeTypes = (resources.statisticalCodeTypes || emptyObj).records || emptyArr;
     const statisticalCodes = (resources.statisticalCodes || emptyObj).records || emptyArr;
     const illPolicies = (resources.illPolicies || emptyObj).records || emptyArr;
@@ -347,7 +363,7 @@ class Instances extends React.Component {
     const callNumberTypes = (resources.callNumberTypes || emptyObj).records || emptyArr;
     const holdingsNoteTypes = (resources.holdingsNoteTypes || emptyObj).records || emptyArr;
     const locations = (resources.locations || emptyObj).records || emptyArr;
-    const locationsById = _.keyBy(locations, 'id');
+    const locationsById = keyBy(locations, 'id');
 
     const referenceTables = {
       contributorTypes,
@@ -361,6 +377,7 @@ class Instances extends React.Component {
       instanceStatuses,
       modesOfIssuance,
       electronicAccessRelationships,
+      instanceNoteTypes,
       statisticalCodeTypes,
       statisticalCodes,
       illPolicies,
@@ -400,7 +417,7 @@ class Instances extends React.Component {
           objectName="inventory"
           maxSortKeys={1}
           searchableIndexes={formattedSearchableIndexes}
-          selectedIndex={_.get(this.props.resources.query, 'qindex')}
+          selectedIndex={get(this.props.resources.query, 'qindex')}
           searchableIndexesPlaceholder={null}
           onChangeIndex={this.onChangeIndex}
           filterConfig={filterConfig}