+++ src/edit/holdings/HoldingsForm.js
@@ -223,41 +223,31 @@ class HoldingsForm extends React.Component {
     /* Menus for Add Item workflow */
     const addHoldingsLastMenu = (
       <PaneMenu>
-        <FormattedMessage id="ui-inventory.createNewHoldingsRecord">
-          {ariaLabel => (
-            <Button
-              buttonStyle="primary paneHeaderNewButton"
-              id="clickable-create-item"
-              type="submit"
-              aria-label={ariaLabel}
-              disabled={(pristine || submitting) && !copy}
-              onClick={handleSubmit(this.onSave)}
-              marginBottom0
-            >
-              <FormattedMessage id="ui-inventory.createHoldingsRecord" />
-            </Button>
-          )}
-        </FormattedMessage>
+        <Button
+          buttonStyle="primary paneHeaderNewButton"
+          id="clickable-create-holdings-record"
+          type="submit"
+          disabled={(pristine || submitting) && !copy}
+          onClick={handleSubmit(this.onSave)}
+          marginBottom0
+        >
+          <FormattedMessage id="stripes-core.button.saveAndClose" />
+        </Button>
       </PaneMenu>
     );
 
     const editHoldingsLastMenu = (
       <PaneMenu>
-        <FormattedMessage id="ui-inventory.updateHoldingsRecord">
-          {ariaLabel => (
-            <Button
-              buttonStyle="primary"
-              id="clickable-update-item"
-              type="submit"
-              aria-label={ariaLabel}
-              disabled={(pristine || submitting) && !copy}
-              onClick={handleSubmit(this.onSave)}
-              marginBottom0
-            >
-              <FormattedMessage id="ui-inventory.updateHoldingsRecord" />
-            </Button>
-          )}
-        </FormattedMessage>
+        <Button
+          buttonStyle="primary"
+          id="clickable-update-item"
+          type="submit"
+          disabled={(pristine || submitting) && !copy}
+          onClick={handleSubmit(this.onSave)}
+          marginBottom0
+        >
+          <FormattedMessage id="stripes-core.button.saveAndClose" />
+        </Button>
       </PaneMenu>
     );