+++ src/edit/items/ItemForm.js
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react';
 import { get, cloneDeep } from 'lodash';
 import PropTypes from 'prop-types';
-import { FormattedMessage } from 'react-intl';
+import { FormattedMessage, injectIntl, intlShape } from 'react-intl';
 import {
   Field,
 } from 'redux-form';
@@ -31,8 +31,8 @@ import {
 } from '@folio/stripes/smart-components';
 
 import RepeatableField from '../../components/RepeatableField';
-
 import ElectronicAccessFields from '../electronicAccessFields';
+import { itemDamageStatuses } from '../../constants';
 
 function validate(values) {
   const errors = {};
@@ -204,6 +204,10 @@ class ItemForm extends React.Component {
 
   onSelectHandler = loc => this.selectTemporaryLocation(loc);
 
+  setItemDamagedStatusDate = () => {
+    this.props.change('itemDamagedStatusDate', new Date());
+  }
+
   render() {
     const {
       handleSubmit,
@@ -224,6 +228,9 @@ class ItemForm extends React.Component {
         statisticalCodeTypes,
       },
       copy,
+      intl: {
+        formatMessage
+      },
     } = this.props;
 
     const {
@@ -233,45 +240,43 @@ class ItemForm extends React.Component {
     } = this.state;
 
     const holdingLocation = locationsById[holdingsRecord.permanentLocationId];
+    const itemDamageOptions = itemDamageStatuses.map(({ label, value }) => (
+      <option
+        value={value}
+        key={value}
+      >
+        {formatMessage({ id: label })}
+      </option>
+    ));
 
     /* Menus for Add Item workflow */
     const addItemLastMenu = (
       <PaneMenu>
-        <FormattedMessage id="ui-inventory.createNewItem">
-          {ariaLabel => (
-            <Button
-              buttonStyle="primary paneHeaderNewButton"
-              id="clickable-create-item"
-              type="submit"
-              aria-label={ariaLabel}
-              disabled={(pristine || submitting) && !copy}
-              onClick={handleSubmit}
-              marginBottom0
-            >
-              <FormattedMessage id="ui-inventory.createItem" />
-            </Button>
-          )}
-        </FormattedMessage>
+        <Button
+          buttonStyle="primary paneHeaderNewButton"
+          id="clickable-create-item"
+          type="submit"
+          disabled={(pristine || submitting) && !copy}
+          onClick={handleSubmit}
+          marginBottom0
+        >
+          <FormattedMessage id="stripes-core.button.saveAndClose" />
+        </Button>
       </PaneMenu>
     );
 
     const editItemLastMenu = (
       <PaneMenu>
-        <FormattedMessage id="ui-inventory.updateItem">
-          {ariaLabel => (
-            <Button
-              buttonStyle="primary"
-              id="clickable-update-item"
-              type="submit"
-              aria-label={ariaLabel}
-              disabled={(pristine || submitting) && !copy}
-              onClick={handleSubmit}
-              marginBottom0
-            >
-              <FormattedMessage id="ui-inventory.updateItem" />
-            </Button>
-          )}
-        </FormattedMessage>
+        <Button
+          buttonStyle="primary"
+          id="clickable-update-item"
+          type="submit"
+          disabled={(pristine || submitting) && !copy}
+          onClick={handleSubmit}
+          marginBottom0
+        >
+          <FormattedMessage id="stripes-core.button.saveAndClose" />
+        </Button>
       </PaneMenu>
     );
 
@@ -681,8 +686,11 @@ class ItemForm extends React.Component {
                         id="input_item_damaged_status_id"
                         component={Select}
                         placeholder={placeholder}
+                        onChange={this.setItemDamagedStatusDate}
                         label={<FormattedMessage id="ui-inventory.itemDamagedStatus" />}
-                      />
+                      >
+                        {itemDamageOptions}
+                      </Field>
                     )}
                   </FormattedMessage>
                 </Col>
@@ -925,6 +933,7 @@ class ItemForm extends React.Component {
 }
 
 ItemForm.propTypes = {
+  intl: intlShape.isRequired,
   onClose: PropTypes.func, // eslint-disable-line react/no-unused-prop-types
   newItem: PropTypes.bool, // eslint-disable-line react/no-unused-prop-types
   handleSubmit: PropTypes.func.isRequired,
@@ -949,4 +958,4 @@ export default stripesForm({
   asyncBlurFields: ['barcode'],
   navigationCheck: true,
   enableReinitialize: true,
-})(ItemForm);
+})(injectIntl(ItemForm));