+++ src/edit/noteFields.js
@@ -5,16 +5,24 @@ import PropTypes from 'prop-types';
 import {
   Icon,
   TextField,
+  Select,
+  Checkbox,
 } from '@folio/stripes/components';
 
 import RepeatableField from '../components/RepeatableField';
 
 const NoteFields = props => {
   const {
+    instanceNoteTypes,
     canAdd,
     canEdit,
     canDelete,
   } = props;
+
+  const instanceNoteTypeOptions = instanceNoteTypes.map(it => ({
+    label: it.name,
+    value: it.id,
+  }));
 
   return (
     <RepeatableField
@@ -26,11 +34,34 @@ const NoteFields = props => {
         </Icon>
       }
       addButtonId="clickable-add-notes"
-      template={[{
-        label: <FormattedMessage id="ui-inventory.note" />,
-        component: TextField,
-        disabled: !canEdit,
-      }]}
+      template={[
+        {
+          name: 'instanceNoteTypeId',
+          label: <FormattedMessage id="ui-inventory.noteType" />,
+          component: Select,
+          disabled: !canEdit,
+          dataOptions: [{ label: 'Select type', value: '' }, ...instanceNoteTypeOptions],
+        },
+        {
+          name: 'note',
+          label: <FormattedMessage id="ui-inventory.note" />,
+          disabled: !canEdit,
+          component: TextField,
+        },
+        {
+          name: 'staffOnly',
+          label: <FormattedMessage id="ui-inventory.staffOnly" />,
+          component: Checkbox,
+          disabled: !canEdit,
+          type: 'checkbox',
+          inline: true,
+          vertical: true,
+          columnSize: {
+            xs: 3,
+            lg: 2,
+          }
+        }
+      ]}
       canAdd={canAdd}
       canDelete={canDelete}
     />
@@ -38,6 +69,7 @@ const NoteFields = props => {
 };
 
 NoteFields.propTypes = {
+  instanceNoteTypes: PropTypes.arrayOf(PropTypes.object),
   canAdd: PropTypes.bool,
   canEdit: PropTypes.bool,
   canDelete: PropTypes.bool,