+++ test/bigtest/interactors/settings/nature-of-content-terms/nature-of-content-terms.js
@@ -1,0 +1,12 @@
+import {
+  interactor,
+  isPresent,
+  count,
+} from '@bigtest/interactor';
+
+@interactor class NatureOfContentTerms {
+  hasList = isPresent('#editList-natureOfContentTerms');
+  rowCount = count('[class^="editListRow---"]');
+}
+
+export default new NatureOfContentTerms();